requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Limiter==3.5.0
Werkzeug==3.0.3
WTForms==3.1.2
SQLAlchemy==2.0.31
Jinja2==3.1.4
python-slugify==8.0.4
Pillow==10.4.0
libsass==0.23.0
python-dotenv==1.0.1
email-validator==2.2.0
bleach==6.1.0
gunicorn==22.0.0
sentry-sdk[flask]==1.40.0
